refactor: minor
