+++ core/src/main/java/com/predic8/membrane/core/interceptor/cors/PreflightHandler.java
@@ -8,11 +8,14 @@ import java.util.*;
 
 import static com.predic8.membrane.core.http.Response.*;
 import static com.predic8.membrane.core.interceptor.Outcome.*;
+import static com.predic8.membrane.core.interceptor.cors.AbstractCORSHandler.ResponseHeaderBuilder.responseBuilder;
 import static com.predic8.membrane.core.interceptor.cors.CorsInterceptor.*;
 import static com.predic8.membrane.core.interceptor.cors.CorsUtil.*;
 import static org.springframework.http.HttpHeaders.*;
 
 public class PreflightHandler extends AbstractCORSHandler {
+
+    private static final Logger log = LoggerFactory.getLogger(PreflightHandler.class);
 
     /**
      * From https://fetch.spec.whatwg.org/#terminology-headers
@@ -24,16 +27,12 @@ public class PreflightHandler extends AbstractCORSHandler {
                 "content-type",
                 "range"
                 );
-    private static final Logger log = LoggerFactory.getLogger(PreflightHandler.class);
 
     public PreflightHandler(CorsInterceptor interceptor) {
         super(interceptor);
     }
 
     public Outcome handleInternal(Exchange exc, String origin) {
-        if (!exc.getRequest().isOPTIONSRequest())
-            return CONTINUE; // no preflight -> let pass
-
         if (interceptor.isAllowAll()) {
             exc.setResponse(noContent().build());
             setCORSHeader(exc, origin);
@@ -74,12 +73,10 @@ public class PreflightHandler extends AbstractCORSHandler {
         if (headers == null)
             return true;
 
-        Set<String> allowedHeaders = interceptor.getAllowedHeaders();
-
         for(String header : toLowerCaseSet(parseCommaOrSpaceSeparated(headers))) {
             if (SAFE_HEADERS.contains(header))
                 continue;
-            if (allowedHeaders.contains(header))
+            if (interceptor.getAllowedHeaders().contains(header))
                 continue;
             log.debug("header '{}' not allowed!", header);
             return false;
@@ -94,5 +91,25 @@ public class PreflightHandler extends AbstractCORSHandler {
 
     protected String getRequestMethod(Exchange exc) {
         return exc.getRequest().getHeader().getFirstValue(ACCESS_CONTROL_REQUEST_METHOD);
+    }
+
+    protected void setCORSHeader(Exchange exc, String requestOrigin) {
+        responseBuilder(exc)
+                .allowOrigin(determineAllowOriginHeader(requestOrigin))
+                .allowMethods(getAllowedMethods(getRequestMethod(exc)))
+                .allowHeaders(getAllowHeaders(getAccessControlRequestHeaderValue(exc)))
+                .maxAge(interceptor.getMaxAge())
+                .allowCredentials(interceptor.getCredentials())
+                .build();
+    }
+
+    private String getAllowHeaders(String requestedHeaders) {
+        if (interceptor.isAllowAll()) {
+            return requestedHeaders != null && !requestedHeaders.isBlank() ? requestedHeaders : "content-type, authorization";
+        }
+        if (!interceptor.getAllowedHeaders().isEmpty()) {
+            return join(List.copyOf(interceptor.getAllowedHeaders()));
+        }
+        return ""; // Todo Check
     }
 }
